Modify the board array to accept the user id as mark
